% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluView-package.R
\docType{package}
\name{FluView-package}
\alias{FluView}
\alias{FluView-package}
\title{FluView: SPC Charts of Flu Data}
\description{
PKG_DESC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/FluView}
  \item Report bugs at \url{https://github.com/d-morrison/FluView/issues}
}

}
\author{
\strong{Maintainer}: Ezra Morrison \email{demorrison@ucdavis.edu}

}
\keyword{internal}
