% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fv}
\alias{fv}
\title{FluView data for California}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 54 rows and 5 columns.
}
\usage{
fv
}
\description{
FluView data for California
}
\keyword{datasets}
