% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presaved_CDC_data}
\alias{presaved_CDC_data}
\title{FluView data for all states}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34074 rows and 19 columns.
}
\usage{
presaved_CDC_data
}
\description{
FluView data for all states
}
\keyword{datasets}
