% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_volume_chart.R
\name{test_volume_chart}
\alias{test_volume_chart}
\title{Title}
\usage{
test_volume_chart(
  dataset = dplyr::mutate(filter(cdcfluview::who_nrevss("state", years =
    2016:2022)$clinical_labs, region == "California"), total_specimens =
    as.numeric(total_specimens))
)
}
\arguments{
\item{dataset}{some subset of `fv`}
}
\value{
a plotly graph
}
\description{
Title
}
