% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv_p_chart.R
\name{fv_p_chart}
\alias{fv_p_chart}
\title{Title}
\usage{
fv_p_chart(
  dataset = dplyr::mutate(filter(presaved_CDC_data, region == "California"),
    total_specimens = as.numeric(total_specimens), `TOTAL POSITIVE` = as.numeric(total_a)
    + as.numeric(total_b)),
  ...
)
}
\arguments{
\item{dataset}{some subset of `fv`}

\item{...}{arguments passed to shewhart.hybrid::PH_Chart(...)}
}
\value{
a plotly graph
}
\description{
Title
}
